+++ test/Algebra/Graph/Test/Labelled/AdjacencyMap.hs
@@ -18,8 +18,8 @@ import Data.Monoid
 
 import Algebra.Graph.Label
 import Algebra.Graph.Labelled.AdjacencyMap
-import Algebra.Graph.Labelled.AdjacencyMap.Internal
 import Algebra.Graph.Test
+import Algebra.Graph.Test.API (toIntAPI, labelledAdjacencyMapAPI)
 import Algebra.Graph.Test.Generic
 import Algebra.Graph.ToGraph (reachable)
 
@@ -27,8 +27,11 @@ import qualified Algebra.Graph.AdjacencyMap as AM
 import qualified Data.Map                   as Map
 import qualified Data.Set                   as Set
 
-t :: Testsuite
-t = testsuite "Labelled.AdjacencyMap." (empty :: LAI)
+tPoly :: Testsuite (AdjacencyMap Any) Ord
+tPoly = ("Labelled.AdjacencyMap.", labelledAdjacencyMapAPI)
+
+t :: TestsuiteInt (AdjacencyMap Any)
+t = fmap toIntAPI tPoly
 
 type S = Sum Int
 type D = Distance Int
@@ -350,7 +353,7 @@ testLabelledAdjacencyMap = do
     test "transpose (edge e x y) == edge e y x" $ \e x y ->
           transpose (edge e x y) == (edge e y x :: LAS)
 
-    test "transpose . transpose == id" $ size10 $ \x ->
+    test "transpose . transpose  == id" $ size10 $ \x ->
          (transpose . transpose) x == (x :: LAS)
 
     putStrLn "\n============ Labelled.AdjacencyMap.gmap ============"